Add files via upload
I figured this could be a good way to approach the game. I made a entities class and ghost/pacman subclasses as well as a location class to store where the entities are but didn't include size information because it seems best to leave them 1x1. 
I added some of the collide method under game logic.
I wasn't sure about the best way to build the board though.